python/requirements.txt
# Application Environment and Parsing
python-dotenv==1.2.1
PyYAML==6.0.3
